test-and-bug.html

<!doctype html>
<html lang="en">

  <head>
    <meta charset="utf-8">

    <title>W3C Test Case and Bug Report</title>

    <meta name="description" content="A session of Test the Web Forward at Shanghai" href="http://testthewebforward.org/events/shanghai-2013.html">
    <meta name="presdate" content="20130818">
    <meta name="author" content="Zhiqiang Zhang" href="mailto:zhiqiang.zhang@intel.com">

    <meta name="apple-mobile-web-app-capable" content="yes" />
    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />

    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">

    <link rel="stylesheet" href="css/reveal.min.css">
    <link rel="stylesheet" href="css/theme/default.css" id="theme">

    <!-- For syntax highlighting -->
    <link rel="stylesheet" href="lib/css/zenburn.css">

    <!-- If the query includes 'print-pdf', use the PDF print sheet -->
    <script>
      document.write( '<link rel="stylesheet" href="css/print/' + ( window.location.search.match( /print-pdf/gi ) ? 'pdf' : 'paper' ) + '.css" type="text/css" media="print">' );
    </script>

    <!--[if lt IE 9]>
    <script src="lib/js/html5shiv.js"></script>
    <![endif]-->
  </head>

  <body>

    <div class="reveal">

      <div class="slides">

        <section>
          <h1>W3C Test Case and Bug Report</h1>
          <p>
            <i>Zhiqiang Zhang (<a href="https://github.com/zqzhang" class="url Nickname">zqzhang</a>), <a class="org" href="http://www.intel.com/">Intel China</a></i>
          </p>
        </section>

        <section>
          <h2>Agenda</h2>
          <ul>
            <li>How to write a spec-based test case?</li>
            <ul>
              <li>Reftest test case</li>
              <li>Script test case</li>
            </ul>
            <li>How to file a good bug report?</li>
          </ul>
        </section>

        <!-- Reftest -->
        <section>
          <section>
            <h2>Reftest Test Case Sample</h2>
            <img src="support/reftest-opera12.png">
            <img src="support/reftest-opera12-ref.png">
          </section>
          <section>
            <h2>Reftest Test Case Sample - Fail</h2>
            <img src="support/reftest-chrome28.png">
            <img src="support/reftest-firefox23.png">
          </section>
          <section>
            <h2>Reftest Test Case Sample - Test Source</h2>
            <pre><code data-trim contenteditable>
&lt;!DOCTYPE html&gt;
&lt;html&gt;
  &lt;head&gt;
    &lt;title&gt;CSS Backgrounds and Borders Test: background-size '52px auto' with background-repeat 'round repeat'&lt;/title&gt;
    &lt;link rel="author" title="Intel" href="http://www.intel.com"&gt;
    &lt;link rel="reviewer" title="G¨¦rard Talbot" href="http://www.gtalbot.org/BrowserBugsSection/css21testsuite/"&gt; &lt;!-- 2012-11-15 --&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-background/#the-background-size" title="3.9. Sizing Images: the 'background-size' property"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-background/#the-background-repeat" title="3.4. Tiling Images: the 'background-repeat' property"&gt;
    &lt;link rel="match" href="reference/background-size-029-ref.xht"&gt;

    &lt;meta name="flags" content="image"&gt;
    &lt;meta name="assert" content="Check if 'background-size' is '52px auto' and 'background-repeat' is 'round repeat', then the width is first rescaled to from 100px to 52px and then rescaled to 60px due to 'round' and the height of the corresponding background image is rescaled from 100px to 60px (to keep the original image aspect ratio) and then repeated vertically."&gt;
    &lt;style&gt;
        div {
            background-color: red;
            background-image: url(support/100x100-blue-and-orange.png);
            background-repeat: round repeat;
            background-size: 52px auto;
            height: 180px;
            width: 180px;
            
            /* 
            Background positioning area is 180px wide by 180px tall.
            The width of the background image is 52px.
            But, because background-repeat is set to round repeat, the
            width is rescaled as following:
            Newest width = 180px / (round [180px / 52px]);
            Newest width = 180px / (round [3.46]);
            Newest width = 180px / (3); 
            Newest width = 60px;                      

            Then the height is rescaled to from 100px to 60px to keep 
            the original aspect ratio.
            */
        }
    &lt;/style&gt;
  &lt;/head&gt;
  &lt;body&gt;
    &lt;p&gt;Test passes if there is 3 rows of 3 blue-and-orange squares and if there is &lt;strong&gt;no partially&lt;/strong&gt; displayed squares and &lt;strong&gt;no red&lt;/strong&gt;.&lt;/p&gt;

    &lt;div&gt;&lt;/div&gt;
  &lt;/body&gt;
&lt;/html&gt;
            </code></pre>
            <p><a href="http://test.csswg.org/source/contributors/intel/submitted/css3-background/background-size-029.html">background-size-029.html</a></p>
          </section>
          <section>
            <h2>Reftest Test Case Sample - Reference Source</h2>
            <pre><code data-trim contenteditable>

&lt;!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd"&gt;

&lt;html xmlns="http://www.w3.org/1999/xhtml"&gt;

 &lt;head&gt;

  &lt;title&gt;CSS Reftest Reference&lt;/title&gt;

  &lt;link rel="author" title="G¨¦rard Talbot" href="http://www.gtalbot.org/BrowserBugsSection/css21testsuite/" /&gt;

  &lt;style type="text/css"&gt;&lt;![CDATA[
  img
  {
  height: 60px;
  vertical-align: top;
  width: 60px;
  }
  ]]&gt;&lt;/style&gt;

 &lt;/head&gt;

 &lt;body&gt;

  &lt;p&gt;Test passes if there is 3 rows of 3 blue-and-orange squares and if there is &lt;strong&gt;no partially&lt;/strong&gt; displayed squares and &lt;strong&gt;no red&lt;/strong&gt;.&lt;/p&gt;

  &lt;div&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;br /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;br /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;img src="../support/100x100-blue-and-orange.png" alt="Image download support must be enabled" /&gt;&lt;/div&gt;

 &lt;/body&gt;
&lt;/html&gt;
            </code></pre>
            <p><a href="http://test.csswg.org/source/contributors/intel/submitted/css3-background/reference/background-size-029-ref.xht">background-size-029-ref.html</a></p>
          </section>
          <section>
            <h2>Reftest Test Case Template</h2>
            <pre><code data-trim contenteditable>
&lt;!DOCTYPE html&gt;
&lt;title&gt;[Test Area]: [Title/Scope of Test]&lt;/title&gt;
&lt;link rel="author" title="[Name of Author]" href=mailto:EMAIL OR 
  http://CONTACT_PAGE"&gt;
&lt;link rel="help" href="http://www.w3.org/TR/[direct link to tested 
  section]"&gt;
&lt;link rel="match" href="[path to reference file]"&gt;
&lt;meta name="flags" content="[requirement flags]"&gt;
&lt;meta name="assert" content="Test checks that [explanation of what 
  you're trying to test]."&gt;
&lt;style&gt;
    [CSS for test]
&lt;/style&gt;
&lt;body&gt;
    &lt;p&gt; Test passes if [description of pass condition].&lt;/p&gt;
    [Content of test]
&lt;/body&gt;
            </code></pre>
            <p><a href="https://github.com/w3c/testtwf-website/blob/gh-pages/docs/test-templates.md#reftest-test-template">test-templates.md#reftest-test-template</a></p>
          </section>
          <section>
            <h2>Reftest Reference Template</h2>
            <pre><code data-trim contenteditable>
&lt;!DOCTYPE html&gt;
&lt;title&gt;[Test Area] Reference File&lt;/title&gt;
&lt;link rel="author" title="[Name of Author]" href=mailto:EMAIL OR 
  http://CONTACT_PAGE"&gt;
&lt;style&gt;
    [CSS for test]
&lt;/style&gt;
&lt;body&gt;
    &lt;p&gt; Test passes if [description of pass condition].&lt;/p&gt;
    [Content of test]
&lt;/body&gt;
            </code></pre>
            <p><a href="https://github.com/w3c/testtwf-website/blob/gh-pages/docs/test-templates.md#reftest-ref-template">test-templates.md#reftest-ref-template</a></p>
          </section>
          <section>
            <h2>Vendor Prefixes</h2>
            <ul>
              <li>You may need to add a vendor prefix to the CSS property you're testing</li>
              <li>If the prefix is necessary, your test will appear to fail without it</li>
              <li>However, do not include vendor prefixes when pushing tests to the W3C</li>
            </ul>
          </section>
        </section>

        <!-- Script test -->
        <section>
          <section>
            <h2>Script Test Case Sample</h2>
            <img src="support/script-test-opera12.png">
            <img src="support/script-test-firefox23.png">
          </section>
          <section>
            <h2>Script Test Case Sample - Test Source</h2>
            <pre><code data-trim contenteditable>
&lt;!DOCTYPE html&gt;
&lt;meta charset=utf-8&gt;
&lt;title&gt;Blob constructor&lt;/title&gt;
&lt;link rel=help href="http://dev.w3.org/2006/webapi/FileAPI/#constructorBlob"&gt;
&lt;link rel=help href="http://dev.w3.org/2006/webapi/WebIDL/#es-union"&gt;
&lt;link rel=help href="http://dev.w3.org/2006/webapi/WebIDL/#es-dictionary"&gt;
&lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
&lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;
&lt;script src="../support/Blob.js"&gt;&lt;/script&gt;
&lt;div id="log"&gt;&lt;/div&gt;
&lt;script&gt;
test(function() {
  assert_true("Blob" in window, "window should have a Blob property.");
  assert_equals(Blob.length, 2, "Blob.length should be 2.");
  assert_true(Blob instanceof Function, "Blob should be a function.");
}, "Blob interface object");

// Step 1.
test(function() {
  var blob = new Blob();
  assert_true(blob instanceof Blob);
  assert_equals(blob.size, 0);
  assert_equals(blob.type, "");
}, "no-argument Blob constructor");
test(function() {
  var blob = Blob();
  assert_true(blob instanceof Blob);
  assert_equals(blob.size, 0);
  assert_equals(blob.type, "");
}, "no-argument Blob constructor without 'new'");

...

&lt;/script&gt;
            </code></pre>
            <p><a href="http://w3c-test.org/web-platform-tests/master/FileAPI/tests/submissions/Ms2ger/Blob-constructor.html">Blob-constructor.html</a></p>
          </section>
          <section>
            <h2>Script Test Template: Shared Metadata</h2>
            <pre><code data-trim contenteditable>
&lt;!DOCTYPE html&gt;
&lt;title&gt;[Test Area]: [Title/Scope of Test]&lt;/title&gt;
&lt;link rel="author" title="[Name of Author]" href=mailto:EMAIL OR 
  http://CONTACT_PAGE"&gt;
&lt;link rel="help" href="http://www.w3.org/TR/[direct link to tested 
  section]"&gt;
&lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
&lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;
&lt;link rel="stylesheet" href="/resources/testharness.css"&gt;
&lt;div id="log"&gt;&lt;/div&gt;
&lt;script&gt;

test(function() {
  [body of test function - must include at least one assert*() 
  function]
}, [Test Name]);

/* Include as many test() functions as you need */

&lt;/script&gt;
            </code></pre>
            <p><a href="https://github.com/w3c/testtwf-website/blob/gh-pages/docs/test-templates.md#script-test-shared-metadata">test-templates.md#script-test-shared-metadata</a></p>
          </section>
          <section>
            <h2>Script Test Template: Per-Test Metadata</h2>
            <pre><code data-trim contenteditable>
&lt;!DOCTYPE html&gt;
&lt;title&gt;[Test Area]&lt;/title&gt;
&lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
&lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;
&lt;link rel="stylesheet" href="/resources/testharness.css"&gt;
&lt;div id="log"&gt;&lt;/div&gt;
&lt;script&gt;

test(function () {
        [body of test function - must include at least one assert*() 
        function]
    }
,  [Test Name]
,   {
        help:   [spec_link_1, 
                 spec_link_2
                ]
    ,   assert: [Test Assertion]
    ,   author: [Author Name &lt;author@email.com]"
    }
);

/* Include as many test() functions as you need */

&lt;/script&gt;
            </code></pre>
            <p><a href="https://github.com/w3c/testtwf-website/blob/gh-pages/docs/test-templates.md#script-test-pertest-metadata">test-templates.md#script-test-pertest-metadata</a></p>
          </section>
        </section>

        <!-- Bug report -->
        <section>
          <section>
            <h2>Bug Report Sample</h2>
            <ul>
              <p>
                <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=43220">Mozilla Bug #43220</a>:
              </p>
              <p>
                Summary:
                <br>
                <small>
                  [CASCADE]author !important rules override user !important rules in user.css
                </small>
                <br>
                Overview:
                <br>
                <small>
                  !important style rules in the author stylesheet override !important style
                  rules in the user stylesheet ([profile]/chrome/user.css); a direct
                  contradiction of the CSS spec (CSS2:6.4)
                </small>
                <br>
                Steps to Reproduce:
                <br>
                <small>
                  1.) Copy 'Sample user.css' to the 'chrome' subdirectory in your profile
                    directory for Mozilla. Rename it to 'user.css'.
                  <br>
                  2.) View 'Testcase' in Mozilla.
                  <br>
                  * Both to be attached shortly
                </small>
                <br>
                Tested with build id=2000061720 on Windows 2000
              </p>
            </ul>
          </section>
          <section>
            <h2>Why Report Bugs?</h2>
            <ul>
              <li>To get the bugs fixed!</li>
              <br>
              <p>Bug Reporting General Guidelines:</p>
              <li>Avoid duplicates: <strong>Search before you file!</strong></li>
              <li><strong>One</strong> bug per report.</li>
              <li>Always test the latest available build.</li>
              <li>State useful facts, not opinions or complaints.</li>
              <li>Flag security/privacy vulnerabilities as non-public.</li>
              <br>
              <li>Follow your project bug reporting guidelines instead of these!</li>
            </ul>
          </section>
          <section>
            <h2>How to Write a Good Bug Report</h2>
            <ul>
              <li>Summary</li>
              <li>Description</li>
              <li>Steps to Reproduce</li>
              <li>Actual Results</li>
              <li>Expected Results</li>
              <li>Reduced Test Case</li>
              <li>Version, Platform and OS</li>
              <li>Crash data and/or Screenshot</li>
            </ul>
          </section>
          <section>
            <h2>Browser Issue Trackers</h2>
            <ul>
              <li>Mozilla's bugzilla <a href="https://bugzilla.mozilla.org/">https://bugzilla.mozilla.org/</a></li>
              <li>WebKit bugs<a href="https://bugs.webkit.org/">https://bugs.webkit.org/</a></li>
              <li>Chrome bugs <a href="https://bugs.webkit.org/">https://bugs.webkit.org/</a></li>
              <li>IE feedback <a href="http://connect.microsoft.com/IE">http://connect.microsoft.com/IE</a></li>
              <br>
              <li>Baidu <a href="http://pcbbs.baidu.com/forum.php?gid=1">http://pcbbs.baidu.com/forum.php?gid=1</a></li>
              <br>
              <li>Tizen <a href="https://bugs.tizen.org/">https://bugs.tizen.org/</a></li>
              <br>
              <li>Other browsers - please check vendor's website for bug reporting information</li>
            </ul>
          </section>
          <section>
            <h2>Specification and Test Case Issue Trackers</h2>
            <ul>
              <li>W3C's bugzilla <a href="https://www.w3.org/Bugs/Public/">https://www.w3.org/Bugs/Public/</a></li>
              <li>W3C mailing list <a href="http://lists.w3.org/">http://lists.w3.org/</a></li>
              <br>
              <li>web-platform-tests issues <a href="https://github.com/w3c/web-platform-tests/issues">https://github.com/w3c/web-platform-tests/issues</a></li>
              <br>
              <li>CSS Test Suite <a href="https://test.csswg.org/shepherd/">https://test.csswg.org/shepherd/</a></li>
            </ul>
          </section>
        </section>

        <section>
          <h2>Reading Source Code!</h2>
          <ul>
            <li><a href="https://github.com/w3c/testtwf-website">https://github.com/w3c/testtwf-website</a></li>
            <br>
            <li><a href="https://github.com/w3c/web-platform-tests">https://github.com/w3c/web-platform-tests</a></li>
            <br>
            <li><a href="https://github.com/w3c/csswg-test">https://github.com/w3c/csswg-test</a></li>
          </ul>
        </section>

        <section>
          <h2>Acknowledgements</h2>
          <ul>
            <li>Rebecca Hauck & Jacob Goldstein, <a href="http://adobe.github.io/web-platform/presentations/test-the-web-forward-tutorial/index.html">Learn how to create a W3C Ref and JS tests</a></li>
            <li>Elika J. Etemad, <a href="http://fantasai.inkedblade.net/style/talks/filing-good-bugs/">How to File a Good Bug Report</a></li>
            <br>
            <li><a href="https://github.com/w3c/testtwf-website/blob/gh-pages/docs/reftests.md">reftests.md</a></li>
            <li><a href="https://github.com/w3c/testtwf-website/blob/gh-pages/docs/testharness-documentation.md">testharness-documentation.md</a></li>
            <li><a href="https://github.com/w3c/testtwf-website/blob/gh-pages/docs/bugreporting.md">bugreporting.md</a></li>
            <br>
            <li>Gerard Talbot, <a href="http://lists.w3.org/Archives/Public/public-css-testsuite/2012Nov/0037.html">background-size-[025-029]: available, improved replacements</a></li>
            <li>Hakim El Hattab, <a href="https://github.com/hakimel/reveal.js">reveal.js</a></li>
          </ul>
        </section>

      </div>

    </div>

    <script src="lib/js/head.min.js"></script>
    <script src="js/reveal.min.js"></script>

    <script>

      // Full list of configuration options available here:
      // https://github.com/hakimel/reveal.js#configuration
      Reveal.initialize({
        controls: true,
        progress: true,
        history: true,
        center: true,

        theme: Reveal.getQueryHash().theme, // available themes are in /css/theme
        transition: Reveal.getQueryHash().transition || 'default', // default/cube/page/concave/zoom/linear/fade/none

        // Optional libraries used to extend on reveal.js
        dependencies: [
          { src: 'lib/js/classList.js', condition: function() { return !document.body.classList; } },
          { src: 'plugin/markdown/marked.js', condition: function() { return !!document.querySelector( '[data-markdown]' ); } },
          { src: 'plugin/markdown/markdown.js', condition: function() { return !!document.querySelector( '[data-markdown]' ); } },
          { src: 'plugin/highlight/highlight.js', async: true, callback: function() { hljs.initHighlightingOnLoad(); } },
          { src: 'plugin/zoom-js/zoom.js', async: true, condition: function() { return !!document.body.classList; } },
          { src: 'plugin/notes/notes.js', async: true, condition: function() { return !!document.body.classList; } }
          // { src: 'plugin/leap/leap.js', async: true }
          // { src: 'plugin/search/search.js', async: true, condition: function() { return !!document.body.classList; } }
          // { src: 'plugin/remotes/remotes.js', async: true, condition: function() { return !!document.body.classList; } }
        ]
      });

    </script>

  </body>
</html>
